backend/requirements.txt
fastapi==0.128.8
uvicorn[standard]==0.30.6
httpx>=0.27.0
anthropic>=0.39.0
python-dotenv>=1.0.0
google-api-python-client>=2.151.0
google-auth>=2.36.0
